feat!: refactor arquitectural completo a hexagonal con migración a Java 25
Se realiza un refactor arquitectural completo del servicio de reportes migrando a una arquitectura hexagonal (Domain-Driven Design) con las siguientes mejoras:

- **Nueva estructura hexagonal**: Organización del código en capas de dominio, aplicación e infraestructura
- **Simplificación de API**: Reemplazo de múltiples clientes Feign por un único InvoiceDataClient que agrega todos los datos de factura
- **Eliminación de Kotlin**: Conversión completa de todos los DTOs y modelos de Kotlin a Java puro
- **Nuevos modelos de dominio**: Creación de modelos Java (Articulo, Cliente, Comprobante, InvoiceData, etc.)
- **Mejora de servicios**: FacturaPdfService ahora acepta directamente InvoiceData para mejor testabilidad
- **Actualización de dependencias**: Spring Boot 4.0.2, Java 25, Spring Cloud 2025.1.0, SpringDoc OpenAPI 3.0.1
- **Pipeline CI/CD actualizado**: Soporte para Java 25 y Eclipse Temurin 25 en Docker
- **Documentación actualizada**: Diagramas Mermaid reflejan la nueva arquitectura hexagonal

BREAKING CHANGE: Se eliminan todos los DTOs Kotlin y los clientes Feign individuales. Los consumidores deben actualizar sus integraciones para usar el nuevo InvoiceDataClient y los modelos Java del paquete domain.model.
