+++ docs/diagrams/mermaid/make-factura-class-diagram.mmd
@@ -8,48 +8,35 @@ classDiagram
 
     class MakeFacturaService {
         -FacturaPdfService facturaPdfService
-        -ClienteMovimientoClient clienteMovimientoClient
+        -InvoiceDataClient invoiceDataClient
         -JavaMailSender javaMailSender
         +send(Long clienteMovimientoId, String email) String
     }
 
     class FacturaPdfService {
         -Environment environment
-        -EmpresaClient empresaClient
-        -ClienteMovimientoClient clienteMovimientoClient
-        -ClienteClient clienteClient
-        -RegistroCaeClient registroCaeClient
-        -MonedaClient monedaClient
-        -ComprobanteClient comprobanteClient
-        -ComprobanteAfipClient comprobanteAfipClient
-        -ArticuloMovimientoClient articuloMovimientoClient
-        -ArticuloClient articuloClient
-        -ConceptoFacturadoClient conceptoFacturadoClient
-        +generatePdf(Long clienteMovimientoId) String
+        -InvoiceDataService invoiceDataService
+        +generatePdf(Long clienteMovimientoId, InvoiceData invoiceData) String
+    }
+
+    class InvoiceDataService {
+        -InvoiceDataClient invoiceDataClient
+        +getInvoiceDataByClienteMovimientoId(Long clienteMovimientoId) InvoiceData
     }
 
     MakeFacturaController --> MakeFacturaService
     MakeFacturaController --> FacturaPdfService
     MakeFacturaService --> FacturaPdfService
-    MakeFacturaService --> ClienteMovimientoClient
-    FacturaPdfService --> EmpresaClient
-    FacturaPdfService --> ClienteMovimientoClient
-    FacturaPdfService --> ClienteClient
-    FacturaPdfService --> RegistroCaeClient
-    FacturaPdfService --> MonedaClient
-    FacturaPdfService --> ComprobanteClient
-    FacturaPdfService --> ComprobanteAfipClient
-    FacturaPdfService --> ArticuloMovimientoClient
-    FacturaPdfService --> ArticuloClient
-    FacturaPdfService --> ConceptoFacturadoClient
+    MakeFacturaService --> InvoiceDataClient
+    FacturaPdfService --> InvoiceDataService
+    InvoiceDataService --> InvoiceDataClient
 
-    interface ClienteMovimientoClient
-    interface EmpresaClient
-    interface ClienteClient
-    interface RegistroCaeClient
-    interface MonedaClient
-    interface ComprobanteClient
-    interface ComprobanteAfipClient
-    interface ArticuloMovimientoClient
-    interface ArticuloClient
-    interface ConceptoFacturadoClient
+    interface InvoiceDataClient {
+        +getInvoiceDataByClienteMovimientoId(Long clienteMovimientoId) InvoiceData
+    }
+
+    class InvoiceData {
+        -ClienteMovimiento clienteMovimiento
+        -RegistroCae registroCae
+        -ClienteMovimiento clienteMovimientoAsociado
+    }
