+++ src/main/java/eterea/report/api/rest/factura/domain/model/Articulo.java
@@ -1,0 +1,10 @@
+package eterea.report.api.rest.factura.domain.model;
+
+import lombok.Getter;
+
+@Getter
+public class Articulo {
+
+    private String descripcion;
+
+}
