+++ src/main/java/eterea/report/api/rest/factura/domain/model/ArticuloMovimiento.java
@@ -1,0 +1,16 @@
+package eterea.report.api.rest.factura.domain.model;
+
+import lombok.Getter;
+
+import java.math.BigDecimal;
+
+@Getter
+public class ArticuloMovimiento {
+
+    private BigDecimal cantidad;
+    private BigDecimal precioUnitarioSinIva;
+    private BigDecimal precioUnitarioConIva;
+    private Articulo articulo;
+    private ConceptoFacturado conceptoFacturado;
+
+}
