+++ src/main/java/eterea/report/api/rest/factura/domain/model/Cliente.java
@@ -1,0 +1,14 @@
+package eterea.report.api.rest.factura.domain.model;
+
+import lombok.Getter;
+
+@Getter
+public class Cliente {
+
+    private String razonSocial;
+    private String domicilio;
+    private String cuit;
+    private Integer posicionIva;
+    private String email;
+
+}
