+++ src/main/java/eterea/report/api/rest/factura/domain/model/ComprobanteAfip.java
@@ -1,0 +1,11 @@
+package eterea.report.api.rest.factura.domain.model;
+
+import lombok.Getter;
+
+@Getter
+public class ComprobanteAfip {
+
+    private Integer comprobanteAfipId;
+    private String label;
+
+}
