+++ src/main/java/eterea/report/api/rest/factura/domain/model/ConceptoFacturado.java
@@ -1,0 +1,10 @@
+package eterea.report.api.rest.factura.domain.model;
+
+import lombok.Getter;
+
+@Getter
+public class ConceptoFacturado {
+
+    private String concepto;
+
+}
