+++ src/main/java/eterea/report/api/rest/factura/domain/model/Empresa.java
@@ -1,0 +1,17 @@
+package eterea.report.api.rest.factura.domain.model;
+
+import lombok.Getter;
+
+@Getter
+public class Empresa {
+
+    private String nombreFantasia;
+    private String razonSocial;
+    private String domicilio;
+    private String telefono;
+    private String condicionIva;
+    private String cuit;
+    private String ingresosBrutos;
+    private String inicioActividades;
+
+}
