+++ src/main/java/eterea/report/api/rest/factura/domain/model/InvoiceData.java
@@ -1,0 +1,12 @@
+package eterea.report.api.rest.factura.domain.model;
+
+import lombok.Getter;
+
+@Getter
+public class InvoiceData {
+
+    ClienteMovimiento clienteMovimiento;
+    RegistroCae registroCae;
+    ClienteMovimiento clienteMovimientoAsociado;
+
+}
