+++ src/main/java/eterea/report/api/rest/factura/domain/ports/in/GetInvoiceDataUseCase.java
@@ -1,0 +1,9 @@
+package eterea.report.api.rest.factura.domain.ports.in;
+
+import eterea.report.api.rest.factura.domain.model.InvoiceData;
+
+public interface GetInvoiceDataUseCase {
+
+    InvoiceData getInvoiceDataByClienteMovimientoId(Long clienteMovimientoId);
+
+}
