--- src/main/java/eterea/report/api/rest/kotlin/model/dto/ComprobanteAfipDto.kt
@@ -1,9 +1,0 @@
-package eterea.report.api.rest.kotlin.model.dto
-
-data class ComprobanteAfipDto(
-
-    var comprobanteAfipId: Int? = null,
-    var nombre: String? = null,
-    var label: String? = null
-
-)
