--- src/main/java/eterea/report/api/rest/kotlin/model/dto/MonedaDto.kt
@@ -1,10 +1,0 @@
-package eterea.report.api.rest.kotlin.model.dto
-
-data class MonedaDto(
-
-    var monedaId: Int? = null,
-    var nombre: String? = "",
-    var simbolo: String? = "",
-    var valorId: Int? = null
-
-)
