+++ src/main/java/eterea/report/api/rest/tool/ToolService.java
@@ -13,9 +13,9 @@ import java.util.regex.Pattern;
 public class ToolService {
 
     public static OffsetDateTime stringDDMMYYYY2OffsetDateTime(String date) {
-        Integer day = Integer.valueOf(date.substring(0, 2));
-        Integer month = Integer.valueOf(date.substring(2, 4));
-        Integer year = Integer.valueOf(date.substring(4, 8));
+        int day = Integer.parseInt(date.substring(0, 2));
+        int month = Integer.parseInt(date.substring(2, 4));
+        int year = Integer.parseInt(date.substring(4, 8));
         return OffsetDateTime.of(year, month, day, 0, 0, 0, 0, ZoneOffset.UTC);
     }
